refactor: create directory for eatch util
