+++ src/utils/addDayDataToChartData/addDayDataToChartData.ts
