+++ src/utils/calculateInitialChartData/calculateInitialChartData.ts
