+++ src/utils/cutFirstRowFirstColumn/cutFirstRowFirstColumn.ts
