--- src/utils/removeLeadingZeros.ts
@@ -1,4 +1,0 @@
-export const removeLeadingZeros = (val: string) => {
-  const leadingZeroes = /^0*/;
-  return val.replace(leadingZeroes, '');
-};
